WIP: Major code refactor
